+++ src/main/java/ru/pavlytskaya/json/TransactionCreatRequest.java
@@ -3,12 +3,11 @@ package ru.pavlytskaya.json;
 import lombok.Data;
 
 import java.math.BigDecimal;
-import java.time.LocalDate;
 
 @Data
 public class TransactionCreatRequest {
-    private Integer accountFrom;
-    private Integer accountTo;
+    private Long accountFrom;
+    private Long accountTo;
     private BigDecimal sum;
     private String data;
 }
